// part9/9.14/src/types.ts
export interface CoursePartBase {
    name: string
    exerciseCount: number
  }

export interface CoursePartWithDescription extends CoursePartBase {
    description: string
}
  
export interface CourseNormalPart extends CoursePartWithDescription {
    type: "normal"
  }

export interface CourseProjectPart extends CoursePartBase {
    type: "groupProject";
    groupProjectCount: number;
  }

export interface CourseSubmissionPart extends CoursePartWithDescription {
    type: "submission"
    exerciseSubmissionLink: string
  }

export interface CourseSpecialPart extends CoursePartWithDescription {
    type: "special"
    requirements: string[]
}
  
export type CoursePart = CourseNormalPart | CourseProjectPart | CourseSubmissionPart | CourseSpecialPart